Refactor: move tooltip documentation to new page and update navigation link
